API_Quanly_Sinh_vien/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
python-dotenv==1.0.0
PyMySQL==1.1.0
cryptography==41.0.7
email-validator==2.1.0
marshmallow==3.20.1
# Testing dependencies
pytest==7.4.3
requests==2.31.0
pytest-cov==4.1.0
gunicorn==21.2.0